Initialize an empty servers_to_monitor to avoid repeat starts
refresh_servers_to_monitor is responsible for detecting added or
removed servers and reacting to them.  It does the following:
1) refresh the servers to monitor list: adding or removing servers from the list.
2) With this server list, start new servers and stop no longer monitored servers.

This logic is used in monitor_servers to ensure the list of monitored servers is
correct and starts or stops based on this list vs. the prior iteration's list.

For start_servers, there is no desired state list we care about, only the
current list of servers from the db, therefore we can avoid using refresh_servers_to_monitor
and update the list directly if it was previously empty.

This avoids having two methods that could start servers in the start_servers
method and possibly starting the same server twice.  This was happening when seeding
the default server AFTER the EvmServer object was initialized, which happens in the start method.