Allow hashes for miq report conditions
In a number of places we pass in hashes for miq expressions.

Report#conditions currently only takes an MiqExpression.
The vast majority of MiqExpressions (if not all) only pass in an expression (i.e.: exp),
which is a simple hash. A hash is a very Yaml friendly data structure.

Not going to change all the reports at this time, but it opens up that possibility
The precedent for this code is MiqAlert#miq_expression=