Remove empty and unused cancel_cleanup method
It's empty and no subclasses implement it.

A search on the manageiq org shows no references to it other than the code
changed in this commit.

It was added in
https://github.com/ManageIQ/manageiq/pull/17825/commits/5f10870c906898e85bff0383a474119f77bff244
but apparantly was never  used.

We can add it back and setting the intermediate state of "canceling" if needed
in the future.  Right now, it's confusing.