Rename supported_subclasses to permitted_subclasses
The concept of "supports" is more centered around specific features
within a provider, however `supported_subclasses` is more centeted on
whether or not specific provider has been permitted or not via the
`Vmdb::PermissionsStores`. This overlap of the term "supported" is
confusing. This commit renames `supported_subclasses` to
`permitted_subclasses` to clarify that distinction.

Additionally it turns out that `supported_types_and_descriptions_hash`
is not used anywhere relevant, so this commit removes that method.