user widgets no longer update group widget order
Before:

When user widgets changed, it updated the group widget ordering.
And widgets were returned from the group that were really part of the user.

After:

Groups are only updated if a widget set belongs to a group

Apology:

Group should never be nil (except for some out of the box dashboards)
I did not look into whether there were issues with the code or in the
factories.