Fix class_for_platform method
The .classify method will, among other things, "un-pluralize". There's
an edge case where the plural form is the proper name - for example
'IbmCloud::PowerVirtualServers'. We used the service's official name,
with the plural ending.

This commit allows the 'platform' to be set to the exact class name
instead of snake case.