Check po files for valid headers
This extends the previous po file spec but checks the header

We are mostly looking to make sure that we didn't accidentally
check in the wrong nplural entry into en/manageiq.po .

The pot file is very similar to english po file, so we often
copy them across. But this makes sure we get the right Language
and plural value.