convert PersistentVolume#parent_name to virtual attribute
virtual_delegate basically buys us all the virtual_attribute goo.
Not sure if the source_type makes this sql friendly, but assuming no.

But it does get us virtual_attribute status, so it will now work in reports

This has been in the report views for a while. not sure how this has been
working up until now.