fix cloud_volume_snapshat_create_queue
When I moved the create snapshot from cloud volume to the
snapshot class, I did not properly move the queuing method across

I had planned on leaving the old queuing method in cloud volume,
so I'm not sure where I screwed up. Regardless, this is calling
a method that does not exist.

This changes the queue method to something that works