Sort gettext output by msgid and sort globbed files for consistency
The intent here is to sort the outputted message catalogs by msgid to hopefully
decrease the line changes each time we run update_all.

Previously, the changes were not random but were not clearly defined and jumped
around the file.  Now, we'll get only line number changes along with actual content
changes, such as if we add or remove features. This should be easier to track than
the prior mechanism which mixed line number changes, content changes and the
same content shifting around in the file due to inconsistent processing
sequence.