Use bootsnap 1.8.1 for psych 4 / 3 compatibility
Psych unsafe_load was added for psych 4 / 3 compatibility but was only available
in 3.3.2 and newer.

Prior to this change, bootsnap 1.8.0 would be break on psych 3.3.1 and older.

Bootsnap added psych 4 compatibility in 1.8.0 by using unsafe_load but this
doesn't work with 3.3.1 or older:
https://github.com/Shopify/bootsnap/pull/368

Bootnap 1.8.1 allows older versions of psych to be compatible via:
https://github.com/Shopify/bootsnap/pull/369