Clear token store in tests
When we override the session store, the TokenStore sometimes holds references
to older sessions.

This seems to happen when we are using non memory session stores:
stub_settings_merge(:server => {:session_store => session_store_value})

We were getting sporadic errors in manageiq-api, and suspect this as the culprit.

This change ensures the session store is cleared out.