Don't memoize a class ivar containing a translation
_() methods are for static strings and will be evaluated by gettext.  Memoizing an
object containing them will lead to a translation occuring the first time
this code is run and in the current locale at that time.  Any changes to locales,
will never cause these strings to be translated again because they're in a memoized
object.

There is no performance reason to memoize here, so it's easy to just remove it.