Remove ANSIBLE_STDOUT_CALLBACK env var
This ENV var is not necessary and causes problems with output with newer
ansible-runner versions such as 2.2.1.

This env var was added in #21382, in order to remove the need for the
ansible.cfg. However, walking further back to when the ansible.cfg was
added leads us to #17564, which was added in order to support
running ansible-playbook (_not ansible-runner_). This variable allows
ansible core to return JSON, which was honored by ansible-playbook.
ansible-runner takes care of wrapping the output properly for us, so we
don't need it, and setting it causes us to lose the "normal" STDOUT we'd
otherwise get from ansible-runner.

ansible-runner 1.4.7 seemingly ignores this value (or perhaps overrides
it for its own purposes), so we don't need to set it.  For the future
versions of ansible-runner, it causes the problems above, so this commit
removes it altogether.