Create MiqRegion for all tests
More and more components of the system are requiring an MiqRegion to function

The straw for me:
I just added a component in MiqQueue that depended upon a value cached in MiqRegion.
Too many tests were failing because there was not a default region.

We did not need to delete all these MiqRegion.seed entries.
I just wanted to delete more lines than I added

In the future, it is possible we will create the region outside the test suite
and remove it from the before(:suite) entry.