Don't automatically create messaging.yml for development
If the messaging.yml file isn't present we assume that messaging/kafka
isn't configured for use and fall back to MiqQueue.

By creating this file with defaults for developers it puts us in a
situation where some functionality will fail if a kafka broker isn't
present on the developer's system.