store/lookup metrics vim_performance_state by correct timestamp
For collecting metrics, we also need to store the vim_performance_state
If there is none available, then we create one for the current time.

This commit is focused on determining if there is already one available.
Since we can only create a vim_performance_state for "NOW", we store
the new record in the cache under "NOW", and we look it up under "NOW"

before
======

lookup a record for a timestamp in the database (but it wasn't in the cache so this will fail)
lookup a record for now in the database - find it

after
=====

lookup a record for now in the cache - find it

We end up with the same data, but since we are looking in the cache for NOW
we are able to detect the records in the cache without 2  db lookups (per record per timestamp)

|        ms |       bytes |    objects |query |   qry ms |     rows |`comments`
|       ---:|         ---:|        ---:|  ---:|      ---:|      ---:| ---
|  18,966.9 | 44,165,289* | 32,711,827 |1,675 |  6,538.2 |      445 |`capture-before`
|  18,789.9 | 62,526,021* | 32,590,959 |1,455 |  6,047.2 |      335 |`capture-after`
|           |             |            | 13%  |      8%  |      25% | delta

Since realtime metrics are a moving target, the number of timestamps is constantly increasing
So it is tricky to properly compare the before and after

But note that we are looking up ~500 fewer VimPerformanceState values. all of those
values are looked up for the current time (across 16 vms and 6 hosts)

This is part of a larger effort. In this example we are collecting in batches
so the numbers may not directly correlate with what is on master, but the savings
will be there anyway

  cluster: 4, hosts: 6, vms: 16