Add VimPerformanceState.capture support for many
We need to lookup VimPerformanceState records for a number of records.

Since we will be creating multiple objects for the same ems / host, we have the potential
to tune this if need be.

We can update the find to lookup states for all objects in a single query.
We can update some of the state capture to leverage duplicates
(at least right now, we're calling for all the same object type, all the same ems,
and probably duplicate hosts.)