Add caching of event_groups and pre-partition the event types
The way this works is we take the resolved set of event_groups and cache
that. This should not change unless the settings change.

After that, instead of always walking the event_groups during lookup, we
walk them once and partition it into one lookup table for String literals
and another lookup table for Regexes, and cache those. Later, when we
want to find the group_and_level we go to those lookup tables instead.

This drops the old lookup code from O(N) + a lot of allocated objects to
O(1 String) + O(N Regex) (note that there are many Strings and very few
Regex).