allow dialogs to be destroyed with read only dialog fields
dialog fields has a read_only field
It was incorrectly assumed to mean seeded in https://github.com/ManageIQ/manageiq/pull/22232
But dialog fields have a number of html attributes, so this was wrong.

Updating the field to allow it to be deleted
Which means that dialogs can be deleted.