Fix chargebacks for Vms with a tenant using a name from settings
We are looking up a tenant by tenant.name
This is a typical technique we use when introducing cross regional support.

Unfortunately, the tenant can have the name overridden by the settings. In this case,
searching by the name would not find the record.

To fix this, we are skipping looking up the tenant when the tenant is in the desired
region.

This will still not work for multi region searching and tenant names are derived from the
settings names.