Fix specs to be relative in time back to avoid changes in the result
The fixed date that's in the specs causes the result to change as time
moves on. In this case, when the PR was written the strings was "about 4
years ago", now it's "over 4 years ago". Instead let's pick a relative
time that will not change over the course of time.

Fixes failed specs after merge of #22426